requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1
django-livereload-server==0.4
django-rest-framework==0.1.0
django-rosetta==0.9.8
djangorestframework==3.14.0
Faker==15.0.0
gunicorn==20.1.0
idna==3.4
livereload==2.6.3
Pillow==9.2.0
pip==22.1.2
polib==1.1.1
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
setuptools==63.4.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tornado==6.2
urllib3==1.26.12
wheel==0.37.1
